+++ CHANGELOG.md
@@ -11,6 +11,18 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 - Populate substitution nodes in postprocess layer (DOCSPLAT-114).
 
+## [v0.3.4] - 2020-02-27
+
+### Removed
+
+- DevHub series directive.
+
+## [v0.3.2] - 2020-02-24
+
+### Fixed
+
+- Devhub directive resolution and behavior.
+
 ## [v0.3.1] - 2020-02-14
 
 ### Added